Print nice error when cloud_provider not supported
The include errors is not user-friendly at all.
Instead, allow the cloud_provider to be wrong in main.yml and check
later that it is supported.

ERROR! An unhandled exception occurred while running the lookup plugin 'first_foun
d'. Error was a <class 'ansible.errors.AnsibleLookupError'>, original message: No
file was found when using first_found. Use errors='ignore' to allow this task to b
e skipped if no files are found